backend/requirements.txt
fastapi
uvicorn[standard]
chess
asyncpg
python-dotenv
